src/main/java/ua/epam/spring/hometask/service/impl/Every10thTicketDiscountStrategy.java
package ua.epam.spring.hometask.service.impl;

import ua.epam.spring.hometask.domain.Event;
import ua.epam.spring.hometask.domain.User;

import javax.annotation.Nonnull;
import javax.annotation.Nullable;
import java.time.LocalDateTime;

public class Every10thTicketDiscountStrategy {
    public byte getDiscount(@Nullable User user, @Nonnull Event event, @Nonnull LocalDateTime airDateTime, long numberOfTickets) {
        if (numberOfTickets % 10 == 0) {
            return 50;
        }

        return 0;
    }
}
